Corrections PR

+++ source/Uniformization/include/Uniformization/Model/Room.h
@@ -27,8 +27,8 @@ namespace adaptone
         const std::vector<Probe>& probes() const;
         const std::vector<Speaker>& speakers() const;
 
-        void setProbesId(const std::vector<size_t>& ids);
-        void setSpeakersId(const std::vector<size_t>& ids);
+        void setProbesId(const std::vector<uint32_t>& ids);
+        void setSpeakersId(const std::vector<uint32_t>& ids);
         void setSpeakerDirectivities(std::size_t speakerIndex, const arma::vec& directivities);
 
         void setProbesPosFromMat(const arma::mat posMat);
